getAddressInfo api parse

+++ app/src/main/java/main/src/etherscan/data/models/TokenBalanceModel.kt
@@ -1,8 +1,9 @@
 package main.src.etherscan.data.models
 
+
 data class TokenBalanceModel(
     val tokenInfo: TokenInfoModel,
-    val balance: Int,
+    val balance: Double,
     val totalIn: Int,
     val totalOut: Int
 )
